Extract ggamp
