+++ packages/core/echo/echo-schema/src/serializer.ts
@@ -10,15 +10,9 @@ import { TextModel } from '@dxos/text-model';
 import { stripUndefinedValues } from '@dxos/util';
 
 import { type EchoDatabase } from './database';
-import {
-  base,
-  type EchoObject,
-  LEGACY_TEXT_TYPE,
-  TextObject,
-  TypedObject,
-  getGlobalAutomergePreference,
-} from './object';
+import { base, type EchoObject, LEGACY_TEXT_TYPE, TextObject, TypedObject } from './object';
 import { Filter } from './query';
+import { getGlobalAutomergePreference } from './automerge-preference';
 
 /**
  * Archive of echo objects.
